1.0.17 RC
Release Candidate:

Removed the stativ samples pattern, now is built dynamically according to the notes in the bank
Activated the sample recording on every note.
At the end of the recording the bank is updated correctly
Notes are immediately available after sampling.
Corrected the bank Json files and removed the notes pattern.